Extract HttpClientHandler from TestExchangeOperator
