+++ presto-main/src/test/java/com/facebook/presto/operator/TestExchangeOperator.java
@@ -13,7 +13,6 @@
  */
 package com.facebook.presto.operator;
 
-import com.facebook.presto.execution.buffer.PagesSerde;
 import com.facebook.presto.execution.buffer.PagesSerdeFactory;
 import com.facebook.presto.execution.buffer.TestingPagesSerdeFactory;
 import com.facebook.presto.metadata.RemoteTransactionHandle;
@@ -23,21 +22,12 @@ import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.split.RemoteSplit;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
-import com.google.common.base.Splitter;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableListMultimap;
-import com.google.common.collect.ImmutableListMultimap.Builder;
-import com.google.common.collect.Iterables;
 import io.airlift.http.client.HttpClient;
-import io.airlift.http.client.HttpStatus;
-import io.airlift.http.client.Request;
-import io.airlift.http.client.Response;
 import io.airlift.http.client.testing.TestingHttpClient;
-import io.airlift.http.client.testing.TestingResponse;
-import io.airlift.slice.DynamicSliceOutput;
 import io.airlift.units.DataSize;
 import io.airlift.units.Duration;
 import org.testng.annotations.AfterClass;
@@ -47,30 +37,21 @@ import org.testng.annotations.Test;
 
 import java.net.URI;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
-import static com.facebook.presto.PrestoMediaTypes.PRESTO_PAGES;
-import static com.facebook.presto.SequencePageBuilder.createSequencePage;
 import static com.facebook.presto.SessionTestUtils.TEST_SESSION;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_BUFFER_COMPLETE;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_PAGE_NEXT_TOKEN;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_PAGE_TOKEN;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_TASK_INSTANCE_ID;
-import static com.facebook.presto.execution.buffer.PagesSerdeUtil.writePages;
-import static com.facebook.presto.execution.buffer.TestingPagesSerdeFactory.testingPagesSerde;
 import static com.facebook.presto.operator.ExchangeOperator.REMOTE_CONNECTOR_ID;
 import static com.facebook.presto.operator.PageAssertions.assertPageEquals;
+import static com.facebook.presto.operator.TestingTaskBuffer.PAGE;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 import static com.facebook.presto.testing.TestingTaskContext.createTaskContext;
 import static io.airlift.concurrent.Threads.daemonThreadsNamed;
 import static io.airlift.units.DataSize.Unit.MEGABYTE;
 import static java.util.concurrent.Executors.newScheduledThreadPool;
-import static javax.ws.rs.core.HttpHeaders.CONTENT_TYPE;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
@@ -79,15 +60,13 @@ import static org.testng.Assert.assertTrue;
 public class TestExchangeOperator
 {
     private static final List<Type> TYPES = ImmutableList.of(VARCHAR);
-    private static final Page PAGE = createSequencePage(TYPES, 10, 100);
     private static final PagesSerdeFactory SERDE_FACTORY = new TestingPagesSerdeFactory();
-    private static final PagesSerde PAGES_SERDE = testingPagesSerde();
 
     private static final String TASK_1_ID = "task1";
     private static final String TASK_2_ID = "task2";
     private static final String TASK_3_ID = "task3";
 
-    private final LoadingCache<String, TaskBuffer> taskBuffers = CacheBuilder.newBuilder().build(CacheLoader.from(TaskBuffer::new));
+    private final LoadingCache<String, TestingTaskBuffer> taskBuffers = CacheBuilder.newBuilder().build(CacheLoader.from(TestingTaskBuffer::new));
 
     private ScheduledExecutorService scheduler;
     private ScheduledExecutorService scheduledExecutor;
@@ -102,7 +81,7 @@ public class TestExchangeOperator
         scheduler = newScheduledThreadPool(4, daemonThreadsNamed("test-%s"));
         scheduledExecutor = newScheduledThreadPool(2, daemonThreadsNamed("test-scheduledExecutor-%s"));
         pageBufferClientCallbackExecutor = Executors.newSingleThreadExecutor();
-        httpClient = new TestingHttpClient(new HttpClientHandler(taskBuffers), scheduler);
+        httpClient = new TestingHttpClient(new TestingExchangeHttpClientHandler(taskBuffers), scheduler);
 
         exchangeClientSupplier = (systemMemoryUsageListener) -> new ExchangeClient(
                 new DataSize(32, MEGABYTE),
@@ -354,89 +333,5 @@ public class TestExchangeOperator
         assertEquals(operator.needsInput(), false);
         assertNull(operator.getOutput());
         assertEquals(operator.getOperatorContext().getOperatorStats().getSystemMemoryReservation().toBytes(), 0);
-    }
-
-    private static class HttpClientHandler
-            implements TestingHttpClient.Processor
-    {
-        private final LoadingCache<String, TaskBuffer> taskBuffers;
-
-        public HttpClientHandler(LoadingCache<String, TaskBuffer> taskBuffers)
-        {
-            this.taskBuffers = taskBuffers;
-        }
-
-        @Override
-        public Response handle(Request request)
-        {
-            ImmutableList<String> parts = ImmutableList.copyOf(Splitter.on("/").omitEmptyStrings().split(request.getUri().getPath()));
-            if (request.getMethod().equals("DELETE")) {
-                assertEquals(parts.size(), 1);
-                return new TestingResponse(HttpStatus.OK, ImmutableListMultimap.of(), new byte[0]);
-            }
-
-            assertEquals(parts.size(), 2);
-            String taskId = parts.get(0);
-            int pageToken = Integer.parseInt(parts.get(1));
-
-            Builder<String, String> headers = ImmutableListMultimap.builder();
-            headers.put(PRESTO_TASK_INSTANCE_ID, "task-instance-id");
-            headers.put(PRESTO_PAGE_TOKEN, String.valueOf(pageToken));
-
-            TaskBuffer taskBuffer = taskBuffers.getUnchecked(taskId);
-            Page page = taskBuffer.getPage(pageToken);
-            headers.put(CONTENT_TYPE, PRESTO_PAGES);
-            if (page != null) {
-                headers.put(PRESTO_PAGE_NEXT_TOKEN, String.valueOf(pageToken + 1));
-                headers.put(PRESTO_BUFFER_COMPLETE, String.valueOf(false));
-                DynamicSliceOutput output = new DynamicSliceOutput(256);
-                writePages(PAGES_SERDE, output, page);
-                return new TestingResponse(HttpStatus.OK, headers.build(), output.slice().getInput());
-            }
-            else if (taskBuffer.isFinished()) {
-                headers.put(PRESTO_PAGE_NEXT_TOKEN, String.valueOf(pageToken));
-                headers.put(PRESTO_BUFFER_COMPLETE, String.valueOf(true));
-                return new TestingResponse(HttpStatus.OK, headers.build(), new byte[0]);
-            }
-            else {
-                headers.put(PRESTO_PAGE_NEXT_TOKEN, String.valueOf(pageToken));
-                headers.put(PRESTO_BUFFER_COMPLETE, String.valueOf(false));
-                return new TestingResponse(HttpStatus.NO_CONTENT, headers.build(), new byte[0]);
-            }
-        }
-    }
-
-    private static class TaskBuffer
-    {
-        private final List<Page> buffer = new ArrayList<>();
-        private int acknowledgedPages;
-        private boolean closed;
-
-        private synchronized void addPages(int pages, boolean close)
-        {
-            addPages(Collections.nCopies(pages, PAGE));
-            if (close) {
-                closed = true;
-            }
-        }
-
-        public synchronized void addPages(Iterable<Page> pages)
-        {
-            Iterables.addAll(buffer, pages);
-        }
-
-        public synchronized Page getPage(int pageSequenceId)
-        {
-            acknowledgedPages = Math.max(acknowledgedPages, pageSequenceId);
-            if (pageSequenceId >= buffer.size()) {
-                return null;
-            }
-            return buffer.get(pageSequenceId);
-        }
-
-        private synchronized boolean isFinished()
-        {
-            return closed && acknowledgedPages == buffer.size();
-        }
     }
 }
